chore: move indicators.js into main.js
